auth and house hooks
