+++ src/components/ImageGallery/ImageGallery.js
@@ -6,8 +6,8 @@ function ImageGallery({images}) {
     <div className='image-gallery'>
        {
          images ?
-         images.map((image) =>
-         <img className='shadow' src={image} width="250px" alt="room-deatil" />
+         images.map((image, i) =>
+         <img key={i} className='shadow' src={image} width="250px" alt="room-deatil" />
          )
          :
          <h5>loading ...</h5>
