+++ src/components/MapComponent/MapComponent.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 function MapComponent() {
     return(
-        <iframe title='map' src="https://www.google.com/maps/embed?pb=!1m14!1m12!1m3!1d63677.41112265173!2d9.715712!3d4.053401599999999!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!5e0!3m2!1sen!2scm!4v1649772269174!5m2!1sen!2scm" width="100%" height="450px" style={{"border":0}} allowfullscreen="" referrerpolicy="no-referrer-when-downgrade"></iframe>
+        <iframe title='map' src="https://www.google.com/maps/embed?pb=!1m14!1m12!1m3!1d63677.41112265173!2d9.715712!3d4.053401599999999!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!5e0!3m2!1sen!2scm!4v1649772269174!5m2!1sen!2scm" width="100%" height="450px" style={{"border":0}} allowFullScreen="" referrerPolicy="no-referrer-when-downgrade"></iframe>
     )
 }
